Возврат на 5-ю версию компилятора, где компилировалось Device

+++ sources/Device/src/Osci/Osci_stm.cpp
@@ -62,7 +62,7 @@ bool Osci::ReadDataChannel(Ch::E ch, uint8 *data)
             {
                 int delta = VALUE::AVE - (int)(*a1);
 
-                int result = (int)(VALUE::AVE - (int)((float)delta * stretch));
+                int result = (int)(VALUE::AVE - (int)(delta * stretch));
 
                 if(result < VALUE::MIN)
                 {
